v0.8.9g

+++ fpf_multinomial.h
@@ -46,8 +46,8 @@ namespace fpf_multinomial {
 			const auto find_multinomial_protein = std::find(par_sample_analysis.multinomial_data.v_protein_name.begin(), par_sample_analysis.multinomial_data.v_protein_name.end(), itr_v_homology_data.blastp_subject_accession);			
 			
 			if ((find_multinomial_element != par_sample_analysis.multinomial_data.v_element_name.end()) && (find_multinomial_protein) != par_sample_analysis.multinomial_data.v_protein_name.end()){
-				const int i = (find_multinomial_element - par_sample_analysis.multinomial_data.v_element_name.begin());
-				const int j = (find_multinomial_protein - par_sample_analysis.multinomial_data.v_protein_name.begin());
+				const size_t i = (find_multinomial_element - par_sample_analysis.multinomial_data.v_element_name.begin());
+				const size_t j = (find_multinomial_protein - par_sample_analysis.multinomial_data.v_protein_name.begin());
 				par_sample_analysis.multinomial_data.v2_frequency[i][j] = itr_v_homology_data.blastp_evalue_transformed_conjugated;
 				par_sample_analysis.multinomial_data.v_frequency_marginal_sum[i] += itr_v_homology_data.blastp_evalue_transformed_conjugated;
 			}
